requirements.txt
flask==2.3.3
langchain==0.1.10
langchain-community==0.0.25
chromadb==0.4.22
python-dotenv==1.0.0
pypdf==6.10.0
python-docx==0.8.11
docx2txt==0.9
sentence-transformers==2.2.2
transformers==4.38.2
torch==2.2.1
requests==2.33.1